Tweaks to fix color coding
Added the block statement for { } back in so properties color code. Not sure why this was removed, so might be overlooking something.

Duplicated keywords in Properties rule set so nested selectors color code. Should be a safe change.

Re-organized the code a little while comparing it to the original CSS version. Cosmetic.